requirements.txt
# Web Framework
fastapi>=0.115.2,<1.0  # Обновлено для совместимости с gradio
uvicorn[standard]==0.24.0
python-multipart>=0.0.18  # Обновлено для совместимости с gradio
starlette>=0.40.0,<1.0  # Обновлено для совместимости с gradio
jinja2==3.1.2

# Video processing
opencv-python==4.12.0.88
imageio==2.31.5
imageio-ffmpeg==0.4.9

# Pose estimation
mediapipe==0.10.14

# Utilities
numpy<2.0  # Совместимость с opencv-python
pillow>=10.0.0

# Deep Learning (GPU)
torch>=2.0.0
torchvision>=0.15.0
# Для CUDA: установите torch с поддержкой CUDA из https://pytorch.org/

# Configuration
pyyaml>=6.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Medical Interface
gradio>=4.0.0

